Make project promotion on home page optional
Add a boolean `promoted` field to the project page model so that the
user can select whether they want to highlight a project in the home
page template tags or not. Only promoted pages will now show up in the
carousel and in the "Current Projects" section of the home page.